requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
Flask==3.0.3
fonttools==4.53.1
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.0.2
openeo==0.31.0
oschmod==0.3.12
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
pystac==1.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
requests==2.32.3
shapely==2.0.6
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
xarray==2024.7.0
zipp==3.20.1
rasterio==1.3.11
folium==0.17.0